requirements.txt
liac-arff 2.5.0
cma 3.3.0
scikit-learn 0.24.2
scikit-multilearn 0.2.0
scipy 1.5.4
numpy 1.19.5
pymoo 0.6.0.1